Updated main page entry to be more reliable
Instead of dispatching a key event to load the main page (can be
unreliable if for whatever reason the login form is not in focus), we
are just going to launch it via javascript command. Also, handled
calling the main page if it is already loaded in the main page (will
happen if the webview session (which persists between application
launches) never logged out).