Action bar implemented, improvements to update process
3 popups can now be launched from the action bar -- change user, update
courselist, and settings. The only one with no functionality at this
point is settings, which at this point is used to set the number of
hours an alarms should be set to before an assignment is due. Also, the
app now detects what the current semester is based on the current date.