A ton of stuff
Axed populating the native android calendar, as it only exists in 4.0
which cuts out a lot of potential users. All alarms are handled by the
app now, using a database to keep track of alarms set so they stay
persistent between app launches. The calendar gui also got an update,
less buggy w/ regards to showing multiple month's dates on a single page
(the last few days and previous few days of the previous month and the
last month, respectively), and also just looks neater. Settings work,
and the user can select to be notified of an upcoming assignment a
varying amount of hours prior.