Major revision first commit
Fixed the update process to not take 3 minutes (I didn't really how http
worked when I first wrote it, my bad). Now it takes around 5-10 seconds
which opens up the gui tremendously. Removed the Calendar in favor for a
List View; the calendar takes too many clicks and is not very user
friendly, so having a list view with one screen of information
definitely helps with that. The UI still needs to be prettied up, but
its at least functional.