Replaced alarm system
The old alarm system was pretty buggy and didn't have a good way to
delete alarms that were already set. Also, the android alarm manager
sets a wakelock for every alarm that gets set, so if dozens of alarms
end up getting set for items months in the future then that can be a
drain of resources (the old alarm system set an alarm for every
assignment in the system). Alarms work now by setting one alarm at a
time (inspiration taken from the android AlarmClock app)