Added lightweight update process
Added a lightweight updater that can take existing course data and then
go directly to that course's portlets for data scraping. This cuts down
the update time from around ~10 seconds to 2-3 seconds. The lightweight
updater uses the user state captured by the last time the "main" updater
was run, so the main update still needs to be run from time to time in
case the user state (i.e. amount of courses) changes.